code clean up

+++ devl/nll/transformation.h
@@ -117,7 +117,7 @@ namespace imaging
          ensure( init.sizex() == 4 && init.sizex() == 4, "only 4x4 matrices are handled" );
          for ( ui32 y = 0; y < 4; ++y )
             for ( ui32 x = 0; x < 4; ++x )
-               _affine( y, x ) = init( y, x );
+               _affine( y, x ) = static_cast<f32>( init( y, x ) );
 
          _affineInverted.clone( _affine );
          bool res = core::inverse( _affineInverted );
